+++ src/index.ts
@@ -1,12 +1,9 @@
-import { Action } from "./action";
 import { Store } from "./store";
-import { Reducer, NamedObservable, StateChangeNotification } from "./types";
+import { Reducer, StateChangeNotification } from "./types";
 import { shallowEqual } from "./shallowEqual"
 export {
-    Action,
     Store,
     Reducer,
-    NamedObservable,
     shallowEqual,
     StateChangeNotification
 }