+++ test/test_action.ts
@@ -1,7 +1,7 @@
-import "mocha";
 import { expect } from "chai";
-
-import { Action, Store, Reducer } from "../src/index";
+import "mocha";
+import { Subject } from "rxjs";
+import { Reducer, Store } from "../src/index";
 
 describe("Action tests", () => {
     interface GenericState {
@@ -10,26 +10,15 @@ describe("Action tests", () => {
     class Foo { };
 
     let store: Store<GenericState>;
-    let genericAction: Action<any>;
+    let genericAction: Subject<any>;
     let genericReducer: Reducer<GenericState, any>;
 
     beforeEach(() => {
         store = Store.create({ value: undefined });
-        genericAction = new Action<any>();
+        genericAction = new Subject<any>();
         genericReducer = (state, payload) => ({ ...state, value: payload });
         store.addReducer(genericAction, genericReducer);
     })
-
-    it("should be possible to create an action without a name", () => {
-        const action = new Action();
-        expect(action.name).to.be.undefined;
-    });
-
-    it("should be possible to create an action with a name", () => {
-        const name = "SOME_NAME";
-        const action = new Action(name);
-        expect(action.name).to.equal(name);
-    });
 
     it("should not throw an error when an action emits a non-plain object", () => {
         expect(() => genericAction.next(new Foo())).not.to.throw();
@@ -43,7 +32,6 @@ describe("Action tests", () => {
     it("should not throw an error when an action emits an array", () => {
         expect(() => genericAction.next([])).not.to.throw();
     });
-
 
     it("should not throw an error when an action emits a string", () => {
         expect(() => genericAction.next("foobar")).not.to.throw();