+++ test/test_reducer.ts
@@ -1,9 +1,8 @@
-import "mocha";
 import { expect } from "chai";
-import { of, range } from "rxjs";
-import { take, toArray, skip } from "rxjs/operators"
-
-import { Action, Reducer, Store } from "../src/index";
+import "mocha";
+import { of, range, Subject } from "rxjs";
+import { skip, take, toArray } from "rxjs/operators";
+import { Reducer, Store } from "../src/index";
 import { ExampleState, SliceState } from "./test_common_types";
 
 describe("Reducer tests", () => {
@@ -23,7 +22,7 @@ describe("Reducer tests", () => {
         // This is a compile time test: we do not want to give a generic type argument to addReducer
         // but compiling with a incompatible reducer will result in compile errors
         // Note: type arguments not expressed on purpose for this test!
-        const addAction = new Action<number>();
+        const addAction = new Subject<number>();
         const addReducer = (state, n) => state + n;
         slice.addReducer(addAction, addReducer);
 
@@ -50,8 +49,20 @@ describe("Reducer tests", () => {
 
     });
 
+    it("should not be possible to pass anything else but observable/string as first argument to addReducer", () => {
+        expect(() => {
+            store.addReducer(5 as any, state => state)
+        }).to.throw();
+    })
+
+    it("should not be possible to pass non-function argument as reducer to addReducer", () => {
+        expect(() => {
+            store.addReducer("foo", 5 as any)
+        }).to.throw();
+    })
+
     it("should not invoke reducers which have been unsubscribed", done => {
-        const incrementAction = new Action<number>();
+        const incrementAction = new Subject<number>();
         const subscription = store.addReducer(incrementAction, (state, payload) => {
             return { ...state, counter: state.counter + payload }
         });
@@ -72,7 +83,7 @@ describe("Reducer tests", () => {
         // This is a compile-time only test to verify the API works nicely.
 
         const incrementReducer: Reducer<number> = (state) => state + 1;
-        const incrementAction = new Action<void>();
+        const incrementAction = new Subject<void>();
         slice.addReducer(incrementAction, incrementReducer);
         slice.select().pipe(skip(1)).subscribe(n => {
             expect(n).to.equal(1);
@@ -96,7 +107,7 @@ describe("Reducer tests", () => {
             const nestedStore = currentStore.createSlice("slice", { foo: "" }) as Store<SliceState>;
 
             const nAsString = n.toString();
-            const fooAction = new Action<string>();
+            const fooAction = new Subject<string>();
             const fooReducer: Reducer<SliceState, string> = (state, payload) => ({ ...state, foo: payload });
             nestedStore.addReducer(fooAction, fooReducer);
             nestedStore.select().pipe(skip(1), take(1)).subscribe(s => {