+++ test/test_stringbased_action_dispatch.ts
@@ -1,6 +1,6 @@
 import "mocha";
 import { expect } from "chai";
-import { Store, Action, Reducer } from "../src/index";
+import { Store, Reducer } from "../src/index";
 import { ExampleState } from "./test_common_types";
 import { Subject } from "rxjs";
 
@@ -43,16 +43,6 @@ describe("String based action dispatch", () => {
 
         it("should be possible to add an action by string and trigger a manual dispatch on it", done => {
             store.addReducer(INCREMENT_ACTION, incrementReducer);
-            store.dispatch(INCREMENT_ACTION, 1)
-            store.select().subscribe(state => {
-                expect(state.counter).to.equal(1);
-                done();
-            })
-        });
-
-        it("should be possible to add an action as NamedObservable and trigger a manual dispatch on it", done => {
-            const incrementAction = new Action<number>(INCREMENT_ACTION)
-            store.addReducer(incrementAction, incrementReducer);
             store.dispatch(INCREMENT_ACTION, 1)
             store.select().subscribe(state => {
                 expect(state.counter).to.equal(1);